refactor(ledger): isolate directory validation
